// apps/portal/src/lib/database/schema.ts
export type Database = {
  public: {
    Tables: {
      keydata: {
        Row: {
          created_at: string
          id: string
          mnemonic: string | null
          user_id: string
        }
        Insert: {
          created_at?: string
          id?: string
          mnemonic?: string | null
          user_id?: string
        }
        Update: {
          created_at?: string
          id?: string
          mnemonic?: string | null
          user_id?: string
        }
        Relationships: [
          {
            foreignKeyName: "keydata_user_id_fkey"
            columns: ["user_id"]
            isOneToOne: false
            referencedRelation: "users"
            referencedColumns: ["id"]
          }
        ]
      }
      projects: {
        Row: {
          created_at: string
          discord: string | null
          domain: string
          google: string | null
          id: string
          name: string
          secret: string
          twitter: string | null
          updated_at: string
          user_id: string
        }
        Insert: {
          created_at?: string
          discord?: string | null
          domain: string
          google?: string | null
          id?: string
          name: string
          secret: string
          twitter?: string | null
          updated_at?: string
          user_id?: string
        }
        Update: {
          created_at?: string
          discord?: string | null
          domain?: string
          google?: string | null
          id?: string
          name?: string
          secret?: string
          twitter?: string | null
          updated_at?: string
          user_id?: string
        }
        Relationships: [
          {
            foreignKeyName: "public_projects_user_id_fkey"
            columns: ["user_id"]
            isOneToOne: false
            referencedRelation: "users"
            referencedColumns: ["id"]
          }
        ]
      }
      users: {
        Row: {
          address: string | null
          created_at: string | null
          id: number
          user_id: string
        }
        Insert: {
          address?: string | null
          created_at?: string | null
          id?: number
          user_id: string
        }
        Update: {
          address?: string | null
          created_at?: string | null
          id?: number
          user_id?: string
        }
        Relationships: [
          {
            foreignKeyName: "users_user_id_fkey"
            columns: ["user_id"]
            isOneToOne: false
            referencedRelation: "users"
            referencedColumns: ["id"]
          }
        ]
      }
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}

export type Tables<
  PublicTableNameOrOptions extends
    | keyof (Database["public"]["Tables"] & Database["public"]["Views"])
    | { schema: keyof Database },
  TableName extends PublicTableNameOrOptions extends { schema: keyof Database }
    ? keyof (Database[PublicTableNameOrOptions["schema"]]["Tables"] &
        Database[PublicTableNameOrOptions["schema"]]["Views"])
    : never = never
> = PublicTableNameOrOptions extends { schema: keyof Database }
  ? (Database[PublicTableNameOrOptions["schema"]]["Tables"] &
      Database[PublicTableNameOrOptions["schema"]]["Views"])[TableName] extends {
      Row: infer R
    }
    ? R
    : never
  : PublicTableNameOrOptions extends keyof (Database["public"]["Tables"] &
      Database["public"]["Views"])
  ? (Database["public"]["Tables"] &
      Database["public"]["Views"])[PublicTableNameOrOptions] extends {
      Row: infer R
    }
    ? R
    : never
  : never
// ...
export type Enums<
  PublicEnumNameOrOptions extends
    | keyof Database["public"]["Enums"]
    | { schema: keyof Database },
  EnumName extends PublicEnumNameOrOptions extends { schema: keyof Database }
    ? keyof Database[PublicEnumNameOrOptions["schema"]]["Enums"]
    : never = never
> = PublicEnumNameOrOptions extends { schema: keyof Database }
  ? Database[PublicEnumNameOrOptions["schema"]]["Enums"][EnumName]
  : PublicEnumNameOrOptions extends keyof Database["public"]["Enums"]
  ? Database["public"]["Enums"][PublicEnumNameOrOptions]
  : never